requirements.txt
google-api-python-client==2.104.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
streamlit==1.39.0
openai>=1.40.0
